% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbour_agg.R
\name{neighbour.agg}
\alias{neighbour.agg}
\title{Neighbouhood index Function - neighbour.agg}
\usage{
neighbour.agg(x, params, wdir, edge.value, logmatrix = F)
}
\arguments{
\item{x}{An input SpatRaster layer which describes the contribution of each cell to fire spread in a "neighbourhood" (e.g., potential Rate of Spread, Head Fire Intensity).}

\item{params}{A dataframe with input parameters, each corresponding to the number of "neighbourhoods". Three parameters required in the dataframe, Distance (param_D), Alpha (param_alpha) and Resolution Factors (param_Rfact).}

\item{wdir}{numeric giving the direction of prevailing wind, with respect to which neighbourhood aggregation should be calculated.}

\item{edge.value}{numeric (or NA) passed to terra's focal function (as fillvalue). The edge.value is used where neighbourhoods extend beyond the raster extent of input x.}

\item{logmatrix}{Logical. Should the moving window(s) generated by the input parameter combination(s) be printed to the log folder? Defaults to F. Log folder will be generated in the working directory.}
}
\description{
This function is an implementation of the Neighbourhood Index for quantifying wildfire likelihood.
For details, please see the publication Radford et al. (2024). An efficient, multi-scale neighbourhood index to quantify wildfire likelihood. The International Journal of Wildland Fire. DOI: 10.1071/WF23055.
}
\examples{
# Load some example input data (e.g., representing Rate of Spread)
  ROS <- rast(system.file("extdata/ROS.tif", package = "Firehoods"))
  
# Define a set of neighbourhood parameters to aggregate the inputs with:
  Neighbourhood_params <- data.frame(param_D = c(500, 2000, 5000, 10000), param_alpha = c(360, 180, 90, 15), param_Rfact = c(2,8,16,32))
  
# Aggregate input data using defined neighbourhoods and plot:
  NI <- neighbour.agg(x = ROS, params = Neighbourhood_params, wdir = 45.0, edge.value = NA, logmatrix = F)
  plot(NI)
}
\keyword{Neighbourhood}
